+++ tests/run-tck-all.py
@@ -34,7 +34,7 @@ SERVERS = {
     'Go':     {'url': 'http://localhost:5000/spec', 'transports': 'jsonrpc,rest'},
     'Python': {'url': 'http://localhost:5000/spec', 'transports': 'jsonrpc,rest'},
     'Rust':   {'url': 'http://localhost:5000/spec', 'transports': 'jsonrpc'},
-    'Java':   {'url': 'http://localhost:5000/spec', 'transports': 'jsonrpc,rest'},
+    'Java':   {'url': 'http://localhost:5000', 'transports': 'jsonrpc,rest'},
 }